move SystemQuery to services.iface package
