+++ java/code/src/com/suse/manager/utils/SaltKeyUtils.java
@@ -19,7 +19,7 @@ import com.redhat.rhn.domain.user.User;
 import com.redhat.rhn.frontend.xmlrpc.PermissionCheckFailureException;
 
 import com.suse.manager.webui.services.impl.SaltService;
-import com.suse.manager.webui.services.impl.SystemQuery;
+import com.suse.manager.webui.services.iface.SystemQuery;
 import com.suse.salt.netapi.calls.wheel.Key;
 
 import java.util.stream.Stream;
