+++ java/code/src/com/suse/manager/webui/controllers/MinionController.java
@@ -47,7 +47,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
-import com.suse.manager.webui.services.impl.SystemQuery;
+import com.suse.manager.webui.services.iface.SystemQuery;
 import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
